python/requirements.txt
numpy
tensorflow-gpu==2.0.0b1
matplotlib
scipy
imageio
scikit-image
pytorch
torchvision
tensorboardX